Refactoring polar_plot.py Extending room parameters with beta
